# Swagger.YAML
swagger: '2.0'
info:
  title: LabMedical API by Geeky Blinders
  version: 1.0.0
...
paths:
...
  /api/exames/{id}:
    parameters:
      - name: id
        in: path
        required: true
        type: string
    put:
      tags:
        - Examinations
      summary: Update Examination by ID
      responses:
        '204':
          description: Examination updated successfully
    get:
      tags:
        - Examinations
      summary: Get Examination by ID
      responses:
        '200':
          description: Successful retrieval of examination by ID
    delete:
      tags:
        - Examinations
      summary: Delete Examination by ID
      responses:
        '204':
          description: Examination deleted successfully
...